Implement building unpacked source in cleanroom
Summary:
This is useful if you want to skip creating a source package, and build
a binary package directly. Useful because it does not require any of the
build dependencies to be installed on the host system -- just a way to
get the source, and Docker.

Test Plan:
Wrote automated tests that run a trivial package through cleanroom. Configured
CircleCI, build here:

https://circleci.com/gh/bitglue/dbuildd/5

Reviewers: #developers, chad

Reviewed By: #developers, chad

Subscribers: chad

Differential Revision: https://phabricator.paperg.com/D325